// src/app/Core/models/user-profile.model.ts
export interface UserProfile {
    id: number;
    firstName: string;
    middleName: string;
    lastName: string;
    efpid: number;
    job: string;
    unit: string;
    extension: number;
    mobile: number;
    email: string;
    accountType: string;
    loginIcon: string;
    username: string;
  }
  